host.attach_static_vdis: disallow cbt_metadata vdi
The impementation calls a script, so it's best to check that none of the
VDIs in the vdi_reason_map are cbt_metadata VDIs. For this API call, we
fail the operation immediately if any of the VDIs have type
cbt_metadata, instead of partially completing it for the
non-cbt_metadata VDIs.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>