VDI.similar_content: don't return cbt_metadata VDIs
Change the VDI.similar_content SMAPIv2 call to not return cbt_metadata
VDIs, as they don't have any actual data. It's safer to do this check
here, as this function is called by storage migration code (from the
Storage_migrate module).

Only drop the cbt_metadata VDIs at the very end of the function, so that
they will be considered when we construct and walk the tree, and parts
of the tree won't be cut off in case we have intermediate cbt_metadata
nodes.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>