Xapi_vm_lifecycle.check_operation_error: filter null VDIs
and invalid VDI references, to improve safety of this function, as it
shouldn't throw an exception - this function is used indirectly in the
message forwarding layer to check the validity of VM operations.
All VDI database accesses should be protected with try/with anyway, as
this function uses a snapshot of the VM record, so the references in the
record might become invalid by the time we use them. This change is to
reduce the likelihood of an exception, in case the try/with is omitted
by accident.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>